// src/db/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  graphql_public: {
    Tables: {
      [_ in never]: never
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      graphql: {
        Args: {
          operationName?: string
          query?: string
          variables?: Json
          extensions?: Json
        }
        Returns: Json
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
  public: {
    Tables: {
      flashcards: {
        Row: {
          back: string
          created_at: string
          front: string
          generation_id: number | null
          id: number
          is_edited: boolean
          source: Database["public"]["Enums"]["flashcard_source"]
          status: Database["public"]["Enums"]["flashcard_status"]
          updated_at: string
          user_id: string
        }
        Insert: {
          back: string
          created_at?: string
          front: string
          generation_id?: number | null
          id?: never
          is_edited?: boolean
          source: Database["public"]["Enums"]["flashcard_source"]
          status?: Database["public"]["Enums"]["flashcard_status"]
          updated_at?: string
          user_id: string
        }
        Update: {
          back?: string
          created_at?: string
          front?: string
          generation_id?: number | null
          id?: never
          is_edited?: boolean
          source?: Database["public"]["Enums"]["flashcard_source"]
          status?: Database["public"]["Enums"]["flashcard_status"]
          updated_at?: string
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "flashcards_generation_id_fkey"
            columns: ["generation_id"]
            isOneToOne: false
            referencedRelation: "generations"
            referencedColumns: ["id"]
          },
        ]
      }
      generations: {
        Row: {
          accepted_edited_count: number
          accepted_unedited_count: number
          created_at: string
          generated_count: number
          generation_duration: number | null
          id: number
          model: string | null
          source_text_hash: string | null
          source_text_length: number
          updated_at: string
          user_id: string
        }
        Insert: {
          accepted_edited_count?: number
          accepted_unedited_count?: number
          created_at?: string
          generated_count: number
          generation_duration?: number | null
          id?: never
          model?: string | null
          source_text_hash?: string | null
          source_text_length: number
          updated_at?: string
          user_id: string
        }
        Update: {
          accepted_edited_count?: number
          accepted_unedited_count?: number
          created_at?: string
          generated_count?: number
          generation_duration?: number | null
          id?: never
          model?: string | null
          source_text_hash?: string | null
          source_text_length?: number
          updated_at?: string
          user_id?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      flashcard_source: "AI" | "MANUAL"
      flashcard_status: "waiting_for_approval" | "approved"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof Database },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends { schema: keyof Database }
  ? Database[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never